fix: スポンサーページのドラフトURLを変更
