+++ app/draft/sponsor/page.tsx
@@ -1,0 +1,9 @@
+import SponsorPage from '@/app/components/SponsorPage'
+
+const DraftSponsorPage = () => {
+  return (
+    <SponsorPage showDraft/>
+  )
+}
+
+export default DraftSponsorPage
